### 2020-07-18 #2491 @labarba
👋  @JamesPaynter — I'm going to send this back for a bit of work on the paper. Currently, it does not follow our guidance in providing a clear _Statement of Need_. See the section [What should my paper contain?](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) in our documentation. Tell us why the research world needs your software!
