### 2020-07-21 #2491 @JamesPaynter
Hi @labarba,
How does whedon identify the statement of need?

I have the lines
"""
`PyGRB` is a pure Python, open source pulse-fitting package which aims bring gamma-ray burst light-curve fitting and analysis into the 21st century.
"""
and
"""
`PyGRB` makes visually appealing and scientifically instructive light-curves from the 4 broadband energy channel BATSE data.
The main feature of `PyGRB` is its ability to fit analytic light-curve models to data.
"""
