+++ joss_paper/paper.bib
@@ -33,12 +33,6 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-
-
-
-
-
-
 @article{astropy,
     author = {{Astropy Collaboration}},
     title = "{Astropy: A community Python package for astronomy}",
