+++ PyGRB/__init__.py
@@ -4,7 +4,7 @@ PyGRB.
 A GRB light-curve analysis package.
 """
 
-__version__ = "0.0.5"
+__version__ = "1.0.1"
 __author__ 	= 'James Paynter'
 
 
