### 2020-07-22 #2491 @JamesPaynter
I have updated the paper (I also fixed a small bug and updated the version to 1.0.1 )
