### 2020-07-24 828ea6b
some minor edits to tte_list class for checking individual detectors. added stte_list to fetch
