### 2020-07-24 #2491 @xuanxu
👋 @fipanther & @zingale - would you be willing to review this submission for JOSS?
I think your knowledge of Python and Astrophysics makes both of you a great fit for this submission.
