### 2020-07-31 #2491 @xuanxu
:wave: @grburgess would you be willing to review this submission for JOSS?
I think your knowledge of Python and Astrophysics makes you a nice fit for this submission.
