### 2020-07-31 #2536 @whedon
[REVIEW]: PyGRB: A pure python gamma-ray burst analysis package.
**Submitting author:** @JamesPaynter (<a href="http://orcid.org/0000-0002-9672-4008">James Paynter</a>)
**Repository:** <a href="https://github.com/JamesPaynter/PyGRB" target ="_blank">https://github.com/JamesPaynter/PyGRB</a>
**Version:** v1.0.2
