### 2020-07-31 #2536 @grburgess
@JamesPaynter A quick run through the code gives the impression that everything is implemented in a statistically proper manner. I have question about how you deal with rates across the BATSE detectors. It appears you are summing them together on a per channel basis, is my impression correct?
