### 2020-08-03 #2536 @JamesPaynter
@grburgess This release (v1.0.1) of the package is aimed at using the pre-binned bfits files that the NASA team creates for each burst after they occur. The available bfits light-curves are generally high enough (time) resolution to perform adequate pulse-fitting. I think they have a slightly more sophisticated algorithm involving the detector response matrices for adding data from multiple detectors, but this is something that I am working through and trying to understand at the moment.

In the cases where I look at the individual photon arrival time data (time tagged event / tte), then I did just naively sum over the detectors. If you know of any literature / tutorials that you could point me towards relating to this, that would greatly help my understanding (My supervisor works in another field and I have been teaching myself - so there are some gaps). Particularly, creating a spectra from stte (spectral) data is currently doing my head in 🤯

But for the main purpose of PyGRB which is pulse-fitting of light-curves, using the pre-binned data is sufficient in ~99% of cases.
