### 2020-08-03 #2536 @zingale
The JOSS guidelines say:

> As a rule of thumb, JOSS’ minimum allowable contribution should represent not less than three months of work for an individual.

the git history shows that the initial commit was ~ 2 months ago.  The SLOC is about 3.5k lines.
