### 2020-08-03 #2536 @JamesPaynter
@zingale I have been working on this for ~2 years. When I was a young and naive programmer I accidentally committed a large amount of files and even after deleting them my repo was ~100mb. So I fiddled around with my git history and eventually wiped it to get the repo to a reasonable size.
