### 2020-08-03 d5e3df4
update installation docs to warn user of bilby warnings
