### 2020-08-03 #2536 @grburgess
@JamesPaynter Thanks for the update! There is actually no way to sum together light curves in count space as the signal from each detector has passed the response and has its own background. In order to model the latent (true) energy-(in)dependent light curve, one of need to fold an energy dependent, spectrally -evolving model through all detectors simultaneously, compare the individual count rates to each of these signals  and sum their log-likelihoods. Otherwise, the result you obtain will have little to do with the incoming signal.

This is an issue with many of the attempts to look for lensed GRBs via detectors that experience a lot of energy dispersion. The observed signal is a very dependent on the local background, incident angle of the GRB, etc. Modeling these effects in the signal space and performing the likelihood in separate data spaces is the only proper way to get around this and will increase the sensitivity of the search. This is however, very computationally expensive. I've written code to do this for GBM, and there are physical background models available (https://www.aanda.org/articles/aa/full_html/2020/08/aa37347-19/aa37347-19.html). The issue of dealing with responses is also highlighted in this article and may give you some insight into how to do this in your code. I'm very happy to help in anyway possible on this problem.

That being said, neglecting the summing of the detectors, everything else seems fine. I have a few more installation checks to do this afternoon, but I would think that your approach is fine on a single detector basis.

It would also help if you could detail in the documentation things like how the background is modeled, what likelihood you are using, etc. I was able to understand that by going through the code, but I knew what I was looking for. A more explicit explanation would help users.
