+++ PyGRB/postprocess/abp.py
@@ -33,7 +33,7 @@ class AbstractBasePlot(object):
             self.plot_dict['p_type']    = 'docs'
             self.plot_dict['width']     = 4
             self.plot_dict['linewidth'] = 0.6
-            self.plot_dict['font_size'] = 14
+            self.plot_dict['font_size'] = 9
             self.plot_dict['heights']  += [0.8] # based off font size
             self.plot_dict['n_axes']    = len(self.plot_dict['heights'])
             self.plot_dict['ext']       = 'png'
@@ -79,78 +79,3 @@ class AbstractBasePlot(object):
             rc('text', usetex=True)
             rc('text.latex',
             preamble=r'\usepackage{amsmath}\usepackage{amssymb}\usepackage{amsfonts}')
-
-    @staticmethod
-    def _make_PP_plot(axes, difference):
-        (osm, osr), (slope, intercept, r) = stats.probplot(difference, dist='norm')
-        axes.scatter(osm, osr, s=0.1, c='k', marker='+')
-        x_vals = np.array([osm[0], osm[-1]])
-        y_vals = intercept + slope * x_vals
-        axes.plot(x_vals, y_vals, 'k--', linewidth=0.3)
-        # adapted from r source code shown in:
-        # https://stats.stackexchange.com/questions/111288/confidence-bands-for-qq-line
-        xx = difference
-        # x<-b0$resid
-        # no nans
-        # good<-!is.na(x)
-        ord = np.sort(xx)
-        # ord<-order(x[good])
-        # masked array
-        # ord.x<-x[good][ord]
-        n = len(xx)
-        # n<-length(ord.x)
-        # ppoints(m, a) = (1:m - a)/(m + (1-a)-a)
-        # 1:m = np.arange(1, m+1), a = 0
-        P = np.arange(1, n + 1) / n
-        # P<-ppoints(n)
-        z = stats.norm.ppf(P)
-        # z<-qnorm(P)
-        # axes_list[2].scatter(z,x)
-        # plot(z,ord.x,type="n")
-        # coeffs =
-        # extracts coefficients after fitting a robust linear model
-        # coef<-coef(rlm(ord.x~z))
-        # a<-coef[1]
-        # b<-coef[2]
-        # abline(a,b,col="red",lwd=2)
-        # confidence interval desired
-        conf = 0.95
-        # conf<-0.95
-        zz = stats.norm.ppf(1 - (1 - conf) / 2)
-        # zz<-qnorm(1-(1-conf)/2)
-        SE = (slope / stats.norm(0, 1).pdf(z)) * np.sqrt(P * (1 - P) / n)
-        # SE<-(b/dnorm(z))*sqrt(P*(1-P)/n)     #[WHY?]
-        fit = intercept + slope * z
-        # fit.value<-a+b*z
-        upper = fit + zz * SE
-        # upper<-fit.value+zz*SE
-        lower = fit - zz * SE
-        # lower<-fit.value-zz*SE
-        axes.plot(z, upper, 'k:', linewidth=0.3)
-        axes.plot(z, lower, 'k:', linewidth=0.3, label='$95\%$ CI')
-        axes.legend()
-
-    @staticmethod
-    def _make_correlogram(axes, x, difference):
-        c, c2 = autocorrelations(difference)
-        # https://online.stat.psu.edu/stat510/lesson/2/2.2
-        # partial acf does not seem to add much more information
-        # from statsmodels.tsa.stattools import acf, pacf
-        # lag_acf = acf(difference)
-        # lag_pacf = pacf(difference)
-        axes.scatter(x[1:], c[1:], s=0.1, c='k', marker='+')
[…]
