### 2020-08-03 28763b6
edit docs for clarity re: products
