### 2020-08-04 #2536 @JamesPaynter
@grburgess That is very interesting! I know you are working mostly with Fermi, but do you know / can you comment if that (spectral-folding) would have been done by the BATSE team during the creation of the bfits light-curves? For example, when I was first playing with the time-tagged photon data for a burst I am interested in, I binned it (summing 3 triggered detectors) with the time bins of the corresponding bfits light-curve. The difference between the two light-curves was 2 counts in about 8,000, so I thought "not bad! this must be how they do it". No doubt there has been a lot of progress in the last 20-30 years though.

And that is the case with present day (Fermi) analysis? Light-curves are created from count data with a spectral model from the get-go? I was wondering recently if it was possible to create a generalised analysis tool that has one model and uses data from various detectors - but it looks like you are already working on something similar in 3ML. I had wanted to simultaneously use data from BATSE's spectral detectors and large area detectors to create a spectrally evolving light-curve fit. But I never came across a spectrally evolving light-curve model in the literature.

I find that extremely interesting, did you publish any results from your gravitational lensing study? I have not seen anything in the literature regarding detector effects, but something I have wondered is how different a GRB could be depending on when / how you observe it. Eg. through the Earth's atmosphere, angle(s) of incidence to detectors, number of triggered detectors etc. To my knowledge all authors so far have assumed the light-curves to be statistically identical within Poisson noise. (Myself included, but only for lack of a more sophisticated modelling capability). It would be an interesting exercise to take a pulse model and see how different you can make at-detector realisations of it.

I read your article cover to cover. It is a really nice analysis and I found it very instructive, particularly the step by step addition of background effects and the discussion of the SAA. I have 15 seconds of 256 channel tte data which I think is similar to Fermi data. The first 500ms contains two pulses which I am alleging are a gravitationally lensed pair. Since the two pulses are so close together in time, the angular response and backgrounds should be the same. A reviewer has (rightly) asked us to compare the spectra of the two pulses. Your paper mentions that the quick and dirty way to create a spectra for transient events is to fit smooth polynomial functions to each channel for off source times, and subtract the background from on source times. I think (hope) this would be sufficient in the above case. The part that I am unsure of is:

1. Combining the data from the three triggered detectors, which from what I now understand will require a spectral model and the detector response matrices(?). Eg. I am not quite sure how to interpret / use Figure 5. in your paper.
2. Dealing with channels with low counts or negative counts after background subtraction.
3. How does a cheat normalise for effective collecting area? Figure 4. of your paper illustrates this quite well, but I think it is a bit more sophisticated than I can achieve here. To first order, is the background uniform? i.e. is the effective area for the background (for normalisation) approximately A_detector, whereas the effective area for the burst is more like A cos theta_burst. Or are the drms really needed to do this?
4. Or do you think I am better off following the model fitting approach you use? In which case, could you please point me towards an appropriate signal model and priors for combining the data across multiple detectors and drms? I could not piece this together from your paper or gbm_drm_gen on GitHub.

I'll make a note in the documentation regarding the (constant) background model and likelihood function. I note your criticism of polynomial background fits in your paper, but I think that for most BATSE bursts polynomial is probably sufficient given the data quality compared to Fermi. (As the lower data quality precludes the scale of background modelling that you achieved with Fermi).

Thanks for your help! Sorry for the essay!
James
