### 2020-08-04 #2536 @JamesPaynter
Looking at your tutorials, I am at the stage of
`# We can check the count spectrum of the detectors, e.g.`
`det_sl['n5'].view_count_spectrum();`
in this tutorial
https://github.com/grburgess/gbm_drm_gen/blob/master/examples/tte_balrog_tutorial.ipynb
with three triggered detectors.
