### 2020-08-04 #2536 @grburgess
So, I am not sure how to divide this up into parts that are constructive for the repository at hand.

You can think of both BATSE and GBM as the same detector technology as both are non-imagining spectrometers suffering from significant energy dispersion. Thus, one must always model in latent space, fold through the detectors, etc to combine information across detectors. If you want to work in a single detector, it is fine to ignore this as long as one realizes what these data actually are.

I do not have a gravitational lens study (I'm rather dubious as to if it is possible) but I do not know of any studies that properly deal with the data. Pulse modeling as you are doing is a good start. Most studies use cross correlation and I think this has a multitude of both model and statistical problems for the data at hand.

I want to emphasize that background is **never** subtracted. It is jointly modeled. Subtraction of Poisson rates leads to all kinds of issues, including when the number of counts is low.

1) Exactly, I have written a tutorial on this here: https://grburgess.github.io/fits_files/ I think you will understand the figure after reading that.

2) There is a myth around low counts that comes from the use of Gaussian likelihoods with Poisson data. Zero counts in a Poisson likelihood is just as informative as 10^3 counts. However, when you write this with the Gaussian "approximation" the MLE estimator for the sigma term gives one the impression that more counts => more information. Your use of a Poisson likelihood is great and appropriate in your code.

3) This is a normal practice in x-ray spectral modeling to deal with total flux calibration. The idea is that there is likely some uncertainty in the responses total effective area. Thus, we can multiply the response (or the predicted counts) of one detector by a constant. This is different than the normalization of the spectral model as it only affects the counts in one detector. This allows for one to correct for unmodeled differences in the total area of the detectors. However, it does not correct for energy dependent issues.

4) I think this goes beyond the scope of the work you have here. I'm not sure about the process (@xuanxu ) and do not want to impede the refereeing for your current code. But I would be happy to have some offline discussions with you and my research group about ways to proceed if you want to go this route.
