### 2020-08-04 #2536 @JamesPaynter
Sorry for going off-topic!
I can shoot you an email  if that is a better place to talk? I would greatly appreciate some help in this regard (point 4). But I will have a look through your tutorial and work over your other suggestions so we can have a more informed discussion.

As for the current code, I am happy with it at the moment. I am just always thinking of improvements.
Thanks again!
