### 2020-08-04 64fd3c8
included expected output and inference specification in docs
