### 2020-08-04 4553979
included expected output and inference specification in docs
