### 2020-08-04 9786cf3
included expected output and inference specification in docs
