+++ docs/source/user/sampling.rst
@@ -37,35 +37,35 @@ Priors
 
 The default priors are
 
-+--------------------------+----------------------------+------------+--------------+--------------+
-| parameter                | minimum                    | maximum    | type         | units        |
-+--------------------------+----------------------------+------------+--------------+--------------+
-| $\\Delta\_i$             | \-\-                       | \-\-       | uniform      | seconds      |
-+--------------------------+----------------------------+------------+--------------+--------------+
-| $\\Delta\_\{i\+1\}$      | $\\Delta\_i$               | \-\-       | uniform      | seconds      |
-+--------------------------+----------------------------+------------+--------------+--------------+
-| $B$                      | $10^\{\-1\}$               | $10^\{3\}$ | log\-uniform | counts / bin |
-+--------------------------+----------------------------+------------+--------------+--------------+
-| $A$                      | $10^\{0\\hphantom\{\-\}\}$ | $10^\{5\}$ | log\-uniform | counts / bin |
-+--------------------------+----------------------------+------------+--------------+--------------+
-| $\\tau$                  | $10^\{\-3\}$               | $10^\{3\}$ | log\-uniform | seconds      |
-+--------------------------+----------------------------+------------+--------------+--------------+
-| $\\xi $                  | $10^\{\-3\}$               | $10^\{3\}$ | log\-uniform | \-\-         |
-+--------------------------+----------------------------+------------+--------------+--------------+
-| $\\gamma$                | $10^\{\-1\}$               | $10^\{1\}$ | log\-uniform | \-\-         |
-+--------------------------+----------------------------+------------+--------------+--------------+
-| $\\nu$                   | $10^\{\-1\}$               | $10^\{1\}$ | log\-uniform | \-\-         |
-+--------------------------+----------------------------+------------+--------------+--------------+
-| $\\Delta\_\\text\{res\}$ | \-\-                       | \-\-       | uniform      | seconds      |
-+--------------------------+----------------------------+------------+--------------+--------------+
-| $A\_\\text\{res\}$       | $10^\{0\\hphantom\{\-\}\}$ | $10^\{3\}$ | log\-uniform | counts / bin |
-+--------------------------+----------------------------+------------+--------------+--------------+
-| $\\tau\_\\text\{res\}$   | $10^\{\-3\}$               | $10^\{3\}$ | log\-uniform | counts / bin |
-+--------------------------+----------------------------+------------+--------------+--------------+
-| $\\omega$                | $10^\{\-3\}$               | $10^\{3\}$ | log\-uniform | \-\-         |
-+--------------------------+----------------------------+------------+--------------+--------------+
-| $\\varphi$               | $\-\\pi$                   | $\\pi$     | uniform      | radians      |
-+--------------------------+----------------------------+------------+--------------+--------------+
++------------------------------------+--------------------------------------+----------------------+--------------+--------------+
+| parameter                          | minimum                              | maximum              | type         | units        |
++------------------------------------+--------------------------------------+----------------------+--------------+--------------+
+| .. math:: $\\Delta\_i$             | \-\-                                 | \-\-                 | uniform      | seconds      |
++------------------------------------+--------------------------------------+----------------------+--------------+--------------+
+| .. math:: $\\Delta\_\{i\+1\}$      | .. math:: $\\Delta\_i$               | \-\-                 | uniform      | seconds      |
++------------------------------------+--------------------------------------+----------------------+--------------+--------------+
+| .. math:: $B$                      | .. math:: $10^\{\-1\}$               | .. math:: $10^\{3\}$ | log\-uniform | counts / bin |
[…]
