### 2020-08-04 3718864
included expected output and inference specification in docs
