+++ docs/source/user/sampling.rst
@@ -37,42 +37,43 @@ Priors
 
 The default priors are
 
-+------------------------------------+--------------------------------------+----------------------+--------------+--------------+
-| parameter                          | minimum                              | maximum              | type         | units        |
-+------------------------------------+--------------------------------------+----------------------+--------------+--------------+
-| .. math:: $\\Delta\_i$             | \-\-                                 | \-\-                 | uniform      | seconds      |
-+------------------------------------+--------------------------------------+----------------------+--------------+--------------+
-| .. math:: $\\Delta\_\{i\+1\}$      | .. math:: $\\Delta\_i$               | \-\-                 | uniform      | seconds      |
-+------------------------------------+--------------------------------------+----------------------+--------------+--------------+
-| .. math:: $B$                      | .. math:: $10^\{\-1\}$               | .. math:: $10^\{3\}$ | log\-uniform | counts / bin |
-+------------------------------------+--------------------------------------+----------------------+--------------+--------------+
-| .. math:: $A$                      | .. math:: $10^\{0\\hphantom\{\-\}\}$ | .. math:: $10^\{5\}$ | log\-uniform | counts / bin |
-+------------------------------------+--------------------------------------+----------------------+--------------+--------------+
-| .. math:: $\\tau$                  | .. math:: $10^\{\-3\}$               | .. math:: $10^\{3\}$ | log\-uniform | seconds      |
-+------------------------------------+--------------------------------------+----------------------+--------------+--------------+
-| .. math:: $\\xi $                  | .. math:: $10^\{\-3\}$               | .. math:: $10^\{3\}$ | log\-uniform | \-\-         |
-+------------------------------------+--------------------------------------+----------------------+--------------+--------------+
-| .. math:: $\\gamma$                | .. math:: $10^\{\-1\}$               | .. math:: $10^\{1\}$ | log\-uniform | \-\-         |
-+------------------------------------+--------------------------------------+----------------------+--------------+--------------+
-| .. math:: $\\nu$                   | .. math:: $10^\{\-1\}$               | .. math:: $10^\{1\}$ | log\-uniform | \-\-         |
-+------------------------------------+--------------------------------------+----------------------+--------------+--------------+
-| .. math:: $\\Delta\_\\text\{res\}$ | \-\-                                 | \-\-                 | uniform      | seconds      |
-+------------------------------------+--------------------------------------+----------------------+--------------+--------------+
-| .. math:: $A\_\\text\{res\}$       | .. math:: $10^\{0\\hphantom\{\-\}\}$ | .. math:: $10^\{3\}$ | log\-uniform | counts / bin |
-+------------------------------------+--------------------------------------+----------------------+--------------+--------------+
-| .. math:: $\\tau\_\\text\{res\}$   | .. math:: $10^\{\-3\}$               | .. math:: $10^\{3\}$ | log\-uniform | counts / bin |
-+------------------------------------+--------------------------------------+----------------------+--------------+--------------+
-| .. math:: $\\omega$                | .. math:: $10^\{\-3\}$               | .. math:: $10^\{3\}$ | log\-uniform | \-\-         |
-+------------------------------------+--------------------------------------+----------------------+--------------+--------------+
-| .. math:: $\\varphi$               | .. math:: $\-\\pi$                   | .. math:: $\\pi$     | uniform      | radians      |
-+------------------------------------+--------------------------------------+----------------------+--------------+--------------+
++----------------------------------+------------------------------------+--------------------+--------------+--------------+
+| parameter                        | minimum                            | maximum            | type         | units        |
++----------------------------------+------------------------------------+--------------------+--------------+--------------+
[…]
