### 2020-08-04 2751cd8
included expected output and inference specification in docs
