### 2020-08-04 54c4462
included expected output and inference specification in docs
