### 2020-08-04 cde771c
included expected output and inference specification in docs
