+++ docs/source/user/sampling.rst
@@ -40,34 +40,34 @@ The default priors are
 +---------------------------------+------------------------------------+--------------------+--------------+--------------+
 | parameter                       | minimum                            | maximum            | type         | units        |
 +---------------------------------+------------------------------------+--------------------+--------------+--------------+
-| .. math:: \Delta_i              | \-\-                               | \-\-               | uniform      | seconds      |
+| :math:`\Delta_i`                | \-\-                               | \-\-               | uniform      | seconds      |
 +---------------------------------+------------------------------------+--------------------+--------------+--------------+
-| .. math:: \Delta_{i\+1}         | .. math:: \Delta\_i                | \-\-               | uniform      | seconds      |
+| :math:`\Delta_{i+1}`            | :math:`\Delta\_i`                  | \-\-               | uniform      | seconds      |
 +---------------------------------+------------------------------------+--------------------+--------------+--------------+
-| .. math:: B                     | .. math:: 10^{\-1}                 | .. math:: 10^{3}   | log\-uniform | counts / bin |
+| :math:`B`                       | :math:`10^{\-1}`                   | :math:`10^{3}`     | log\-uniform | counts / bin |
 +---------------------------------+------------------------------------+--------------------+--------------+--------------+
-| .. math:: A                     | .. math:: 10^{0\hphantom{\-}}      | .. math:: 10^{5}   | log\-uniform | counts / bin |
+| :math:`A`                       | :math:`10^{0}`                     | :math:`10^{5}`     | log\-uniform | counts / bin |
 +---------------------------------+------------------------------------+--------------------+--------------+--------------+
-| .. math:: \tau                  | .. math:: 10^{\-3}                 | .. math:: 10^{3}   | log\-uniform | seconds      |
+| :math:`\tau`                    | :math:`10^{-3}`                    | :math:`10^{3}`     | log\-uniform | seconds      |
 +---------------------------------+------------------------------------+--------------------+--------------+--------------+
-| .. math:: \xi                   | .. math:: 10^{\-3}                 | .. math:: 10^{3}   | log\-uniform | \-\-         |
+| :math:`\xi`                     | :math:`10^{-3}`                    | :math:`10^{3}`     | log\-uniform | \-\-         |
 +---------------------------------+------------------------------------+--------------------+--------------+--------------+
-| .. math:: \gamma                | .. math:: 10^{\-1}                 | .. math:: 10^{1}   | log\-uniform | \-\-         |
+| :math:`\gamma`                  | :math:`10^{-1}`                    | :math:`10^{1}`     | log\-uniform | \-\-         |
 +---------------------------------+------------------------------------+--------------------+--------------+--------------+
-| .. math:: \nu                   | .. math:: 10^{\-1}                 | .. math:: 10^{1}   | log\-uniform | \-\-         |
+| :math:`\nu`                     | :math:`10^{-1}`                    | :math:`10^{1}`     | log\-uniform | \-\-         |
 +---------------------------------+------------------------------------+--------------------+--------------+--------------+
-| .. math:: \Delta_\text{res}     | \-\-                               | \-\-               | uniform      | seconds      |
+| :math:`\Delta_\text{res}`       | \-\-                               | \-\-               | uniform      | seconds      |
 +---------------------------------+------------------------------------+--------------------+--------------+--------------+
-| .. math:: A_\text{res}          | .. math:: 10^{0\hphantom{\-}}      | .. math:: 10^{3}   | log\-uniform | counts / bin |
+| :math:`A_\text{res}`            | :math:`10^{0}`                     | :math:`10^{3}`     | log\-uniform | counts / bin |
[…]
