### 2020-08-04 #2536 @zingale
I'm happy with this once the contributor guidelines are updated (issue tagged above ^).

I'll leave the technical details about working with GRB data to the other reviewer, as that is not my expertise.
