+++ docs/source/user/sampling.rst
@@ -44,7 +44,7 @@ The default priors are
 +---------------------------------+------------------------------------+--------------------+--------------+--------------+
 | :math:`\Delta_{i+1}`            | :math:`\Delta\_i`                  | \-\-               | uniform      | seconds      |
 +---------------------------------+------------------------------------+--------------------+--------------+--------------+
-| :math:`B`                       | :math:`10^{\-1}`                   | :math:`10^{3}`     | log\-uniform | counts / bin |
+| :math:`B`                       | :math:`10^{-1}`                    | :math:`10^{3}`     | log\-uniform | counts / bin |
 +---------------------------------+------------------------------------+--------------------+--------------+--------------+
 | :math:`A`                       | :math:`10^{0}`                     | :math:`10^{5}`     | log\-uniform | counts / bin |
 +---------------------------------+------------------------------------+--------------------+--------------+--------------+
@@ -60,14 +60,14 @@ The default priors are
 +---------------------------------+------------------------------------+--------------------+--------------+--------------+
 | :math:`A_\text{res}`            | :math:`10^{0}`                     | :math:`10^{3}`     | log\-uniform | counts / bin |
 +---------------------------------+------------------------------------+--------------------+--------------+--------------+
-| :math:`\tau_\text{res}`         | :math:`10^{\-3}`                   | :math:`10^{3}`     | log\-uniform | counts / bin |
+| :math:`\tau_\text{res}`         | :math:`10^{-3}`                    | :math:`10^{3}`     | log\-uniform | counts / bin |
 +---------------------------------+------------------------------------+--------------------+--------------+--------------+
-| :math:`\omega`                  | :math:`10^{\-3}`                   | :math:`10^{3}`     | log\-uniform | \-\-         |
+| :math:`\omega`                  | :math:`10^{-3}`                    | :math:`10^{3}`     | log\-uniform | \-\-         |
 +---------------------------------+------------------------------------+--------------------+--------------+--------------+
 | :math:`\varphi`                 | :math:`-\pi`                       | :math:`\pi`        | uniform      | radians      |
 +---------------------------------+------------------------------------+--------------------+--------------+--------------+
 
-The priors on :math:`\Delta_i` are determined by the length of the light-curve passed to `PulseFitter`.
+The priors on :math:`\Delta_i` and :math:`\Delta_\text{res}` are determined by the length of the light-curve passed to `PulseFitter`.
 
 
 Further reading
