### 2020-08-05 80f88fb
edit sampling docs
