+++ docs/source/user/sampling.rst
@@ -1,3 +1,4 @@
+.. _sampling:
 Sampling
 ========
 
@@ -17,19 +18,41 @@ The equation should read
     S(t|A,\Delta,\tau,\xi) = A \exp \left[ - \xi \left(  \frac{t - \Delta}{\tau} + \frac{\tau}{t-\Delta}  \right)  -2 \right]
 
 
+Poisson rate
+------------
 
+The rate passed to the likelihood function is the sum of the individual pulses, specified in :ref:`_pulses`.
+An example rate for two FRED pulses would be:
+
+.. math::
+
+    S(t|A_1,\Delta_1,\tau_1,\xi_1,A_2,\Delta_2,\tau_2,\xi_2) =
+                      A_1 \exp \left[ - \xi_1 \left(  \frac{t - \Delta_1}{\tau_1}
+                              + \frac{\tau_1}{t-\Delta_1}  \right)  -2 \right]
+                    + A_2 \exp \left[ - \xi_2 \left(  \frac{t - \Delta_2}{\tau_2}
+                              + \frac{\tau_2}{t-\Delta_2}  \right)  -2 \right]
 
 Background
-----------
+^^^^^^^^^^
 
 The background is by default modelled as a constant.
 This is sufficient for all but the longest gamma-ray bursts, excepting periods of exceptional background variation.
 More complex background models, such as a polynomial, can be included by specifying them as a rate function and including relevant priors.
 
+The complete rate is then:
+
+.. math::
+
+    S(t|B,A_1,\Delta_1,\tau_1,\xi_1,A_2,\Delta_2,\tau_2,\xi_2) = B +
+                      A_1 \exp \left[ - \xi_1 \left(  \frac{t - \Delta_1}{\tau_1}
+                              + \frac{\tau_1}{t-\Delta_1}  \right)  -2 \right]
+                    + A_2 \exp \left[ - \xi_2 \left(  \frac{t - \Delta_2}{\tau_2}
+                              + \frac{\tau_2}{t-\Delta_2}  \right)  -2 \right]
+
 Likelihood
 ----------
 
-The likelihood function is a Poisson likelihood, which is a sum of the specified rates.
+The rate function is then passed into the Poisson likelihood, which is a sum of the specified rates.
 
 
 Priors
