### 2020-08-05 757374b
fixed math overflow
