### 2020-08-05 5d3d4d8
fixed math overflow
