+++ docs/source/user/sampling.rst
@@ -29,7 +29,7 @@ An example rate for two FRED pulses would be:
     \begin{split}
     S(t|A_1,\Delta_1,\tau_1,\xi_1,A_2,\Delta_2,\tau_2,\xi_2) =
                       &A_1 \exp \left[ - \xi_1 \left(  \frac{t - \Delta_1}{\tau_1}
-                              + \frac{\tau_1}{t-\Delta_1}  \right)  -2 \right]
+                              + \frac{\tau_1}{t-\Delta_1}  \right)  -2 \right] \\
                     + &A_2 \exp \left[ - \xi_2 \left(  \frac{t - \Delta_2}{\tau_2}
                               + \frac{\tau_2}{t-\Delta_2}  \right)  -2 \right]
     \end{split}
@@ -47,9 +47,9 @@ The complete rate is then:
 .. math::
 
   \begin{split}
-  S(t|A_1,\Delta_1,\tau_1,\xi_1,A_2,\Delta_2,\tau_2,\xi_2) = B + 
+  S(t|A_1,\Delta_1,\tau_1,\xi_1,A_2,\Delta_2,\tau_2,\xi_2) = B +
                     &A_1 \exp \left[ - \xi_1 \left(  \frac{t - \Delta_1}{\tau_1}
-                            + \frac{\tau_1}{t-\Delta_1}  \right)  -2 \right]
+                            + \frac{\tau_1}{t-\Delta_1}  \right)  -2 \right] \\
                   + &A_2 \exp \left[ - \xi_2 \left(  \frac{t - \Delta_2}{\tau_2}
                             + \frac{\tau_2}{t-\Delta_2}  \right)  -2 \right]
   \end{split}
