### 2020-08-05 #2536 @JamesPaynter
@grburgess "It would also help if you could detail in the documentation things like how the background is modeled, what likelihood you are using, etc"

I updated the documentation (priors, background, likelihood, rates), but I wasn't sure how much information is encapsulated by the 'etc'.
https://pygrb.readthedocs.io/en/latest/user/sampling.html
