### 2020-08-05 #2536 @zingale
I didn't test the automated tests, but I see there is a travis file, and a test coverage link and tests/ directory.  So I am happy with that.
