### 2020-08-06 545c24f
new branch for BATSE preprocess restructure
