+++ PyGRB/preprocess/BATSE/counts/basecounts.py
@@ -1,0 +1,152 @@
+"""
+A preprocessing module to unpack the BATSE tte and discsc bfits FITS files.
+Written by James Paynter, 2020.
+"""
+
+import numpy as np
+import pandas as pd
+import matplotlib.pyplot as plt
+from astropy.io import fits
+from pathlib import Path
+
+
+from PyGRB.fetch.get_BATSE import GetBATSEBurst
+
+class BaseBATSE(object):
+    """ A base class for BATSE rate / count data and detector response matrices.
+    """
+
+    def __init__(self, trigger, datatype, **kwargs):
+        super(BaseBATSE, self).__init__()
+        self.trigger = trigger
+        self.datatype = datatype
+        self.times = kwargs.pop('times', 'full')
+
+        fetch = GetBATSEBurst(trigger = self.trigger, datatype = self.datatype, **kwargs)
+        with fits.open(fetch.path) as hdu_list:
+            self.header           = hdu_list[0].header
+            self.calibration_data = hdu_list[1].data
+            try:
+                # try to set count data
+                self.count_data       = hdu_list[2].data
+            except:
+                # drms only have one data HDU.
+                pass
+
+    def _get_energy_bin_edges(self):
+        """ Get the energy bin edges from the FITS file.
+            Will have dimensions of (nEnergy_bins + 1, nDetectors).
+        """
+        self.energy_bin_edges = self.calibration_data['E_EDGES']
+
+    def print_headers(self):
+        """ Print the header info from the FITS file. """
+        print('Header Data Unit Info')
+        print('---------------')
+        print(self.header)
+        print('---------------')
+        print('Header Data Unit Calibration Info')
+        print('---------------')
+        print(self.calibration_data.columns)
+        print('---------------')
+        try:
+            print('Header Data Unit Calibration Info')
+            print('---------------')
+            print(self.count_data.columns)
+            print('---------------')
+        except:
+            # drms only have one data HDU.
+            pass
+
+class BaseBurstBATSE(BaseBATSE):
+    """ A base class for BATSE burst data. """
+
+    def __init__(self, *args, **kwargs):
+        super(BaseBurstBATSE, self).__init__(*args, **kwargs)
+
+        self.rates      = self.count_data['RATES']
+        self.bin_left   = self.count_data['TIMES'][:,0]
+        self.bin_right  = self.count_data['TIMES'][:,1]
+        self.bin_widths = self.bin_right - self.bin_left
+
+        (self.nBins, self.nChannels) = np.shape(self.rates)
+        self.channels = np.arange(self.nChannels)
+
+        self._get_time_edges()
+
+    def _get_time_edges(self):
+        """ Define the start and stop time of the burst. """
+        try:
+            (self.t_start, self.t_stop) = self.times
+        except:
+            if self.times == 'T90':
+                self._read_T90_table()
+            elif self.times == 'T100':
+                self._read_T90_table()
+                self.t_start = min(-2, self.t_start)
+                self.t_stop += max(5, 0.25 * self.t90)
+            elif self.times == 'full':
+                self.t_start, self.t_stop = self.bin_left[0], self.bin_right[-1]
+
+    def _open_T90_excel(self):
+        """ Open the BATSE 4B csv information file. """
+        xls_file = f'../../../data/BATSE_4B_catalogue.xls'
+        path = Path(__file__).parent / xls_file
+        cols = ['trigger_num', 't90', 't90_error', 't90_start']
+        dtypes = {  'trigger_num': np.int32, 't90' : np.float64,
+                    't90_error' : np.float64, 't90_start' : np.float64}
+        try:
+            table = pd.read_excel(path, sheet_name = 'batsegrb', header = 0,
[…]
