### 2020-08-07 9318487
working on stte spectra with fitspec.py
