+++ docs/source/user/sampling.rst
@@ -70,7 +70,7 @@ The default priors are
 +---------------------------------+------------------------------------+--------------------+--------------+--------------+
 | :math:`\Delta_i`                | \-\-                               | \-\-               | uniform      | seconds      |
 +---------------------------------+------------------------------------+--------------------+--------------+--------------+
-| :math:`\Delta_{i+1}`            | :math:`\Delta\_i`                  | \-\-               | uniform      | seconds      |
+| :math:`\Delta_{i+1}`            | :math:`\Delta_i`                  | \-\-               | uniform      | seconds      |
 +---------------------------------+------------------------------------+--------------------+--------------+--------------+
 | :math:`B`                       | :math:`10^{-1}`                    | :math:`10^{3}`     | log\-uniform | counts / bin |
 +---------------------------------+------------------------------------+--------------------+--------------+--------------+

[…]
