+++ README.rst
@@ -1,7 +1,6 @@
 |JOSS| |Travis| |Coverage| |Docs| |pypi| |version| |AstroPy|
 
-.. figure:: https://github.com/JamesPaynter/PyGRB/blob/master/docs/source/images/logo.png
-    :width: 100%
+.. image:: https://github.com/JamesPaynter/PyGRB/blob/master/docs/source/images/logo.png
     :align: center
     :alt: PyGRB logo
 
@@ -56,7 +55,6 @@ Pulse types
 Description of GRB pulse phenomenology.
 
 .. image:: https://github.com/JamesPaynter/PyGRB/blob/master/docs/source/images/equations/FRED.gif
-    :width: 100%
     :align: center
     :alt: FRED eqn: $I(t) = A \exp{ - \xi \left( \frac{t - \Delta}{\tau} + \frac{\tau}{t-\Delta} \right)}$
 
@@ -71,7 +69,6 @@ Description of GRB pulse phenomenology.
    :language: python
 
 .. image:: https://github.com/JamesPaynter/PyGRB/blob/master/docs/source/images/BATSE_trigger_7475_rates_rates.png
-    :width: 80%
     :align: center
     :alt: BATSE trigger 7475
 
@@ -138,7 +135,6 @@ The data products are stored in :code:`products/`.
 We should be left with a light-curve that looks like this:
 
 .. image:: https://github.com/JamesPaynter/PyGRB/blob/master/docs/source/images/B_7475__d_NL200__rates_F.png
-    :width: 80%
     :align: center
     :alt: BATSE trigger 7475
 
