+++ PyGRB/preprocess/BATSE/satellite/ibdb.py
@@ -1,0 +1,23 @@
+""" Python script for reading and interpreting BATSE ibdb.fits files.
+"""
+
+
+class BaseIBDB(object):
+    """
+    A class for reading and interpreting BATSE ibdb.fits files.
+
+    """
+
+    def __init__(self):
+        super(BaseIBDB, self).__init__()
+
+
+    @staticmethod
+    def _bit_to_byte():
+        pass
+
+    def read_science_data(self):
+        pass
+
+if __name__ == '__main__':
+    pass
