+++ PyGRB/main/fitspec.py
@@ -41,7 +41,7 @@ class SpecFitter(object):
         stte = SpectralTTE(3770)
         stte._get_energy_bin_edges()
 
-        det = 5
+        det = 7
         stte.live_detectors = [det]
         drm = drms.drm[f'Detector {det}']
 
@@ -50,18 +50,95 @@ class SpecFitter(object):
 
 
         pm = PhotonMask(detector = det,
-                        # channel_low = 4, channel_high = 255,
+                        channel_low = 4, channel_high = 255,
                         time_start = 1, time_stop = 15,
                         # energy_low = 10,
                         )
 
         photons = pm.get_masked_photons(stte.photons)
 
-        u, counts = stte.get_counts_per_channel(photons)
+        u, counts   = stte.get_counts_per_channel(photons)
         energy_bins = np.sort(np.unique(photons[:,5]))
+        channels    = np.sort(np.unique(photons[:,2].astype('int'))) - 4
+        # print(channels)
 
+
+        channnnels = np.sort(np.unique(stte.photons[:,2]))
+        print(len(channnnels))
+
+
+
+
+        # e_low  = np.sort(np.unique(photons[:,3]))
+        # e_high = np.sort(np.unique(photons[:,4]))
+
+        # PHA channels from drm
+        e_low  = drms.energy_channels[f'Detector {det}'][:-1] # len 258
+        e_high = drms.energy_channels[f'Detector {det}'][1:]  # len 258
+        e_mid = (e_high + e_low) / 2
+        # integrate PHA channels
+        true_fluxes = integral(e_low, e_high) # len 258
+
+        # fig, ax = plt.subplots()
+        # ax2 = ax.twinx()
+        # ax.bar(e_low, height = true_fluxes, width = e_high - e_low,
+        #         align = 'edge', label = 'model', alpha = 0.5, fill = 'False',
+        #         facecolor = 'None', edgecolor = 'k')
+        # ax2.scatter(e_mid, e_high - e_low, marker = '.', label = 'ebins')
+        # ax.set_xscale('log')
+        # ax.set_yscale('log')
+        # ax.set_ylabel(r'Differential Photon Flux $\frac{d N_p}{dt dA}$')
+        # ax2.set_yscale('log')
+        # ax2.set_ylabel('Energy Bin Width')
+        # plt.legend()
+        # # plt.show()
+        print(np.shape(drm)) # (252, 258), 252 real channels, 258 PHA channels
+
+        # convolve true_fluxes with drm to get dN/dt
+        folded_counts = np.dot(true_fluxes, drm.T) # len 252 normalises by area
+
+        dE = np.diff(drms.energy_bins[f'Detector {det}']) # len 252
+        norm_folded_counts = folded_counts / dE # len 252
+        # norm_counts = counts / dE
+
+
+        def folded_power_law(energy_bins, norm, index):
+
+            def _power_law(energy, norm, index):
+                return norm * np.power(energy/100.,index)
+
+            def _differential_flux(e):
+                return power_law(e, .01, -2)
+
+            # integral of the differential flux
+            def _integral(e1, e2):
+                return (e2 - e1) / 6.0 * (_differential_flux(e1)
+                + 4 * _differential_flux((e1 + e2) / 2.0)+ _differential_flux(e2))
+
+            e_high = energy_bins[1:]
+            e_low  = energy_bins[:-1]
+            true_fluxes = _integral(e_low, e_high)
+            # print(len(true_fluxes))
+            dot = np.dot(true_fluxes, drm[:,channels].T) / dE
+            # print('drm', np.shape(drm[:,channels]))
+            # print('dot', len(dot))
+            # print('dot', dot[channels])
[…]
