+++ PyGRB/main/fitspec.py
@@ -28,7 +28,8 @@ def integral(e1, e2):
     return (e2 - e1) / 6.0 * (differential_flux(e1)
     + 4 * differential_flux((e1 + e2) / 2.0)+ differential_flux(e2))
 
-
+def log_mid(low, high):
+    return (low + high) / 2
 
 class SpecFitter(object):
     """docstring for SpecFitter."""
@@ -41,7 +42,7 @@ class SpecFitter(object):
         stte = SpectralTTE(3770)
         stte._get_energy_bin_edges()
 
-        det = 7
+        det = 5
         stte.live_detectors = [det]
         drm = drms.drm[f'Detector {det}']
 
@@ -50,55 +51,47 @@ class SpecFitter(object):
 
 
         pm = PhotonMask(detector = det,
-                        channel_low = 4, channel_high = 255,
+                        channel_low = 13, channel_high = 255,
                         time_start = 1, time_stop = 15,
                         # energy_low = 10,
                         )
 
         photons = pm.get_masked_photons(stte.photons)
-
-        u, counts   = stte.get_counts_per_channel(photons)
-        energy_bins = np.sort(np.unique(photons[:,5]))
-        channels    = np.sort(np.unique(photons[:,2].astype('int'))) - 4
-        # print(channels)
-
-
-        channnnels = np.sort(np.unique(stte.photons[:,2]))
-        print(len(channnnels))
-
-
-
-
-        # e_low  = np.sort(np.unique(photons[:,3]))
-        # e_high = np.sort(np.unique(photons[:,4]))
-
+        # PHA channels from stte file
+        photon_channels, photon_counts = stte.get_counts_per_channel(photons) # len = 246
+        photon_energy_bins = np.sort(np.unique(photons[:,5]))
         # PHA channels from drm
-        e_low  = drms.energy_channels[f'Detector {det}'][:-1] # len 258
-        e_high = drms.energy_channels[f'Detector {det}'][1:]  # len 258
-        e_mid = (e_high + e_low) / 2
+        drm_pha_channel_lo = drms.PHA_channel_edges[f'Detector {det}'][:-1] # len 252
+        drm_pha_channel_hi = drms.PHA_channel_edges[f'Detector {det}'][1:]  # len 252
+        drm_pha_channel_mid= log_mid(drm_pha_channel_lo, drm_pha_channel_hi)
         # integrate PHA channels
-        true_fluxes = integral(e_low, e_high) # len 258
+        true_fluxes = integral(drm_pha_channel_lo, drm_pha_channel_hi) # len 252
 
-        # fig, ax = plt.subplots()
-        # ax2 = ax.twinx()
-        # ax.bar(e_low, height = true_fluxes, width = e_high - e_low,
-        #         align = 'edge', label = 'model', alpha = 0.5, fill = 'False',
-        #         facecolor = 'None', edgecolor = 'k')
-        # ax2.scatter(e_mid, e_high - e_low, marker = '.', label = 'ebins')
-        # ax.set_xscale('log')
-        # ax.set_yscale('log')
-        # ax.set_ylabel(r'Differential Photon Flux $\frac{d N_p}{dt dA}$')
-        # ax2.set_yscale('log')
-        # ax2.set_ylabel('Energy Bin Width')
-        # plt.legend()
-        # # plt.show()
-        print(np.shape(drm)) # (252, 258), 252 real channels, 258 PHA channels
+        drm_energy_bins_lo = drms.incident_photon_bins[f'Detector {det}'][:-1] # len 258
+        drm_energy_bins_hi = drms.incident_photon_bins[f'Detector {det}'][1:]  # len 258
[…]
