### 2020-08-25 e5980c7
Update gravlensnature.py
