### 2020-08-25 a176c36
Update makepriors.py
