+++ extra/hardnessduration.py
@@ -1,106 +1,168 @@
 import numpy as np
 import matplotlib.pyplot as plt
 
-# from PyGRB.preprocess.BATSE.counts.ttebfits
+from PyGRB.preprocess.BATSE.counts.ttebfits import TimeTaggedRates
 
-duration_table = 'extra/duration_table_5b.txt'
-with open(duration_table) as file:
-    lines = file.readlines()
+GRB = TimeTaggedRates(3770)
+bg = GRB._get_rough_backgrounds() * 0.005
+times_A = GRB.bin_left[(GRB.bin_left > -0.1) & (GRB.bin_left < 0.15)]
+counts_A = GRB.rates[  (GRB.bin_left > -0.1) & (GRB.bin_left < 0.15)] * 0.005
 
+times_B = GRB.bin_left[(GRB.bin_left > 0.3) & (GRB.bin_left < 0.55)]
+counts_B = GRB.rates[  (GRB.bin_left > 0.3) & (GRB.bin_left < 0.55)] * 0.005
 
-drn_arr = np.zeros((10000,24))
-#  0 trigger number
-#  1 T50
-#  2 T50 err
-#  3 T50 start
-#  4 T90
-#  5 T90 err
-#  6 T90 start
-#  7 fluence in channel 1
-#  8 fluence in channel 1 error
-#  9 fluence in channel 2
-# 10 fluence in channel 2 error
-# 11 fluence in channel 3
-# 12 fluence in channel 3 error
-# 13 fluence in channel 4
-# 14 fluence in channel 4 error
-# 15 peak flux in 64ms timescale
-# 16 peak flux in 64ms timescale error
-# 17 peak flux in 64ms time
-# 18 peak flux in 256ms timescale
-# 19 peak flux in 256ms timescale error
-# 20 peak flux in 256ms time
-# 21 peak flux in 1024ms timescale
-# 22 peak flux in 1024ms timescale error
-# 23 peak flux in 1024ms time
+times_S = GRB.bin_left[(GRB.bin_left > -0.1) & (GRB.bin_left < 0.55)]
+counts_S = GRB.rates[  (GRB.bin_left > -0.1) & (GRB.bin_left < 0.55)] * 0.005
 
-for line in lines:
-    line = line.strip()
-    words = line.split(' ')
+for i, c in enumerate(['r', 'orange', 'g', 'b']):
+    plt.step(times_A, counts_A[:,i], c = c)
+    plt.step(times_B, counts_B[:,i], c = c)
 
-    data = []
-    for word in words:
-        if len(word) > 0:
-            data.append(word)
-    drn_arr[int(data[0]),:7] = np.array(data).astype('float')
+def counts_AA(i):
+    return np.sum(counts_A[:,i] - bg[i])
+
+def counts_BB(i):
+    return np.sum(counts_B[:,i] - bg[i])
+
+def counts_SS(i):
+    return np.sum(counts_S[:,i] - bg[i])
+
+hard_A = counts_AA(2) / counts_AA(1)
+hard_B = counts_BB(2) / counts_BB(1)
+hard_S = counts_SS(2) / counts_SS(1)
+hard_A_err = np.sqrt(1/counts_AA(2)+1/counts_AA(1)) * hard_A
+hard_B_err = np.sqrt(1/counts_BB(2)+1/counts_BB(1)) * hard_B
+hard_S_err = np.sqrt(1/counts_SS(2)+1/counts_SS(1)) * hard_S
+f_A = f'The hardness of pulse A is {hard_A} $\pm$ {hard_A_err}'
+f_B = f'The hardness of pulse B is {hard_B} $\pm$ {hard_B_err}'
[…]
