### 2020-09-02 ce41c60
PyGRB\preprocess\GRB_class.py
