+++ extra/hardnessduration.py
@@ -1,58 +1,167 @@
 import numpy as np
 import matplotlib.pyplot as plt
+from matplotlib import rc
 
 from PyGRB.preprocess.BATSE.counts.ttebfits import TimeTaggedRates
 
-GRB = TimeTaggedRates(3770)
-bg = GRB._get_rough_backgrounds() * 0.005
-times_A = GRB.bin_left[(GRB.bin_left > -0.1) & (GRB.bin_left < 0.15)]
-counts_A = GRB.rates[  (GRB.bin_left > -0.1) & (GRB.bin_left < 0.15)] * 0.005
+fig, ax = plt.subplots(nrows = 1)
+T90s = np.zeros(8122)
+H32s = np.zeros(8122)
+bursts = np.array([105,107,108])
+# for i in bursts:
+# for i in range(105, 8122):
+#     try:
+#         GRB = TimeTaggedRates(i, times = 'T90')
+#         if not np.isnan(GRB.t90):
+#             T90s[i] = GRB.t90
+#             t = GRB.bin_left
+#             s = GRB.t_start
+#             bg = GRB._get_rough_backgrounds()
+#             r2 = GRB.rates[:,1] - bg[1]
+#             r3 = GRB.rates[:,2] - bg[2]
+#             H2 = np.sum(r2[(t > s) & (t < (s + GRB.t90))])
+#             H3 = np.sum(r3[(t > s) & (t < (s + GRB.t90))])
+#             # print('t' = t)
+#             print('i = ', i)
+#             # print('start =', s)
+#             # print('T90 =', s + GRB.t90)
+#             # print('H2 = ', H2)
+#             # print('H3 = ', H3)
+#             # print('\n\n')
+#             H32s[i] = H3 / H2
+#     except:
+#         pass
 
-times_B = GRB.bin_left[(GRB.bin_left > 0.3) & (GRB.bin_left < 0.55)]
-counts_B = GRB.rates[  (GRB.bin_left > 0.3) & (GRB.bin_left < 0.55)] * 0.005
+# mT = np.ma.masked_where(T90s==0, T90s)
+# mH = np.ma.masked_where(T90s==0, H32s)
+# masked_T90 = np.ma.compressed(mT)
+# masked_H32 = np.ma.compressed(mH)
 
-times_S = GRB.bin_left[(GRB.bin_left > -0.1) & (GRB.bin_left < 0.55)]
-counts_S = GRB.rates[  (GRB.bin_left > -0.1) & (GRB.bin_left < 0.55)] * 0.005
+# with open('masked_T90.txt', 'w') as f:
+#     for item in masked_T90:
+#         f.write("%s\n" % item)
+# with open('masked_H32.txt', 'w') as f:
+#     for item in masked_H32:
+#         f.write("%s\n" % item)
+with open('masked_T90.txt', 'r') as f1:
+    masked_T90_l = np.array(f1.read().split('\n'))
+with open('masked_H32.txt', 'r') as f2:
+    masked_H32_l = np.array(f2.read().split('\n'))
+masked_T90 = np.zeros(len(masked_T90_l))
+masked_H32 = np.zeros(len(masked_H32_l))
+for i in range(len(masked_H32_l)):
+    try:
+        masked_T90[i] = float(masked_T90_l[i])
+        masked_H32[i] = float(masked_H32_l[i])
+    except:
+        pass
 
[…]
