### 2020-09-08 #2536 @JamesPaynter
@xuanxu I think there is just one last issue to close off re: instrumental dead time / using proper count data.
I must admit that I have been meaning to chase this up with @grburgess but I got distracted by another paper's referee feedback that I am addressing.
