### 2020-09-08 #2536 @grburgess
Yes, I was just waiting to see what the plans were to deal with a few of the issues I raised. Especially re: summing of detectors.
