### 2020-09-08 #2536 @JamesPaynter
Should we continue this discussion on my GitHub issues?

re: detector summing, is it then appropriate to have a radiation field which is convolved / folded with the triggered detector response matrices to then compare to observed counts in the Poisson likelihood?
This would work for the tte data.

I'm still not sure of a way forward using the main available datatypes (eg discsc 64ms binned data) as my understanding is they come pre-summed over triggered detectors (prebinned on the satellite). There is not tte data available to cover the entire time history of ~80-90% of BATSE GRBs, only lower resolution eg. discsc data.
