### 2020-09-10 06a2215
about to put in luminsoity function calculation
