+++ paper/lensing.py
@@ -1,0 +1,329 @@
+import numpy as np
+import matplotlib.pyplot as plt
+
+import scipy.integrate as integrate
+from scipy.optimize import fsolve
+import math
+
+import matplotlib as mpl
+from matplotlib import rc
+mpl.rcParams['text.latex.preamble'] = [r'\usepackage{amsmath}']
+rc('font',**{'family':'DejaVu Sans','serif':['Computer Modern'], 'size':12})
+rc('text', usetex=True)
+
+from scipy.stats import poisson
+
+cols = ['tab:blue', 'tab:orange', 'tab:green', 'tab:red', 'tab:purple']
+
+dlambda = 1e-6
+llambda = np.linspace(dlambda, 10, int(10/dlambda))
+p = poisson.pmf(1, llambda)
+prior = 1 / np.sqrt(llambda)
+p2 = prior * p
+norm_p = p2 / np.sum(p2 * dlambda)
+
+fig, ax = plt.subplots(figsize = (8,5))
+ax.plot(llambda, norm_p, c = 'k')
+
+threshold = 8e-6/8e-5
+zeds = [0.05, 0.5, 1, 2, 5]
+lums = [1e43, 1e44, 1e45, 1e46, 1e47]
+omegas = [.333, 3.22e-3, 5.4e-5, 3e-6, 1e-6] #lum
+omegas = [1.8e-2, 3e-4, 1.5e-4, 8e-5, 6e-5] # z l = 1e45
+# omegas = [1.8e-2, 3e-4, 1.5e-4, 8e-5, 6e-5] #z l = 1e44.5
+for i, (z, omega, l) in enumerate(zip(zeds, omegas, lums)):
+    threshold = 8e-6 / omega
+
+    credibility = np.sum(dlambda * p2[llambda < threshold])
+    print(f'At <L> = {l} a globular cluster is excluded at {100*(1-credibility):.3f}% credibility')
+    # print(f'At <z_s> = {z} a globular cluster is excluded at {100*(1-credibility):.3f}% credibility')
+    ax.axvline(threshold, color = cols[i], linestyle = ':',
+        label = f'$\\Omega_\\text{{gc}}/\\Omega_l$ for $\\left<z_s\\right>\\sim{z}$')
+ax.set_xscale('log')
+ax.set_xlabel(r'$X\sim$ Poisson$(N=1)$')
+ax.set_ylabel('Probability Density')
+ax.legend(loc = 'upper left')
+# plt.show()
+
+const_G = 6.6741e-11 # SI units
+const_c = 299792458.  # m/s
+omega_m = 0.286
+omega_v = 0.714
+H_0     = 69600. #/ 3.086e22    # m/s / Mpc
+MPC_2_M = 3.086e22
+
+GRB_LUMINOSITY = 10**44.5 # Watts
+ENERGY_TO_GAMMA = 1.602e-14 # Watts to 100 kev phot / sec
+PHI_0 = 0.3 # photons / cm^2 / sec (BATSE minimum detectable photon flux)
+LENS_MASS = 6e4 * 2e30 # kg
+MIN_TIME_DELAY = 1e-2 # second
+MAX_TIME_DELAY = 240 # second
+SOURCE_REDSHIFT = 5
+
+# not used
+# def pd(z_l, z_s):
+#     return (vector_pd(z_s) - vector_pd(z_l)) / (1 + z_s)
+def efficiency(phi):
+    phi_0 = 0.21  # photons / sec / cm^2
+    alpha = 0.105 # photons / sec / cm^2
+    return np.where(phi < phi_0, 0, 1 - np.exp(- (phi - phi_0) / alpha))
+def cross_section(mass, z_l, z_s):
[…]
