### 2020-09-14 #2536 @grburgess
Once a caveat about the summing of detectors is either stated in the docs or fixed in the code, I'm ok with this being published. Nice work @JamesPaynter .
