### 2020-09-15 898ca2f
update sampling docs
