### 2020-09-15 1e29d5a
update sampling docs
