### 2020-09-15 aa1795d
update sampling docs
