### 2020-09-15 0796321
update sampling docs
