+++ docs/source/refs.bib
@@ -1,6 +1,33 @@
-@Book{1987:nelson,
-  author = {Edward Nelson},
-  title = {Radically Elementary Probability Theory},
-  publisher = {Princeton University Press},
-  year = {1987}
+@ARTICLE{2008GeoRLGrefenstette,
+       author = {{Grefenstette}, B.~W. and {Smith}, D.~M. and {Dwyer}, J.~R. and
+         {Fishman}, G.~J.},
+        title = "{Time evolution of terrestrial gamma ray flashes}",
+      journal = "{Geophysical Research Letters}",
+     keywords = {Atmospheric Processes: Atmospheric electricity, Atmospheric Processes: Lightning, Atmospheric Processes: Instruments and techniques},
+         year = 2008,
+        month = mar,
+       volume = {35},
+       number = {6},
+          eid = {L06802},
+        pages = {L06802},
+          doi = {10.1029/2007GL032922},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2008GeoRL..35.6802G},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{2010JGRAGjesteland,
+       author = {{Gjesteland}, T. and {{\O}stgaard}, N. and {Connell}, P.~H. and
+         {Stadsnes}, J. and {Fishman}, G.~J.},
+        title = "{Effects of dead time losses on terrestrial gamma ray flash measurements with the Burst and Transient Source Experiment}",
+      journal = {Journal of Geophysical Research (Space Physics)},
+     keywords = {Atmospheric Processes: Atmospheric electricity, Mathematical Geophysics: Instruments and techniques, Mathematical Geophysics: Spectral analysis (3205, 3280), Atmospheric Composition and Structure: Instruments and techniques},
+         year = 2010,
+        month = may,
+       volume = {115},
+       number = {3},
+          eid = {A00E21},
+        pages = {A00E21},
+          doi = {10.1029/2009JA014578},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2010JGRA..115.0E21G},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
