+++ docs/source/user/pulses.rst
@@ -1,4 +1,5 @@
 .. _pulses:
+
 Pulse Types
 ===========
 
