### 2020-09-15 9855e7b
finished sampling docs
