### 2020-09-15 #2536 @JamesPaynter
Thank you @grburgess 😄

I updated the docs in the sampling section to reflect this:
https://pygrb.readthedocs.io/en/latest/user/sampling.html
