### 2020-09-15 9bde1e0
finished sampling docs
