### 2020-09-15 #2536 @xuanxu
OK @JamesPaynter, everything looks good now, here are the next steps:

- Please release a new [tagged version](https://github.com/JamesPaynter/PyGRB/releases) from the current main branch so it includes all the changes made during the review process
- Then archive that latest release in Zenodo (or a similar service).
- Check the Zenodo deposit has the correct metadata: title and author name should match the paper; you may also add author ORCID.

Once you do that please report here the version number and archive DOI
