### 2020-09-15 69c0b73
Introduce GRB
Introduce the GRB acronym used later in the paper

Re: https://github.com/openjournals/joss-reviews/issues/2536
