+++ .readthedocs.yml
@@ -12,4 +12,4 @@ sphinx:
 python:
   version: 3.7
   install:
-    - pip install sphinxcontrib-bibtex
+    - requirements: docs/requirements.txt
