### 2020-09-15 7dc3b75
Update sampling.rst
