### 2020-09-15 #2536 @JamesPaynter
Thanks! :)

New version v1.0.2

DOI https://doi.org/10.5281/zenodo.4031843

I linked my ORCID id to the Zenodo account, is there more to do to link it to the JOSS paper?

https://orcid.org/0000-0002-9672-4008
