### 2020-09-16 #2536 @xuanxu
>  is there more to do to link it to the JOSS paper?

It's already in the [metadata of the paper](https://github.com/JamesPaynter/PyGRB/blob/master/joss_paper/paper.md)
