### 2020-09-16 #2536 @JamesPaynter
Merged and updated in Zenodo :)
