### 2020-09-16 #2536 @kyleniemeyer
Hi @JamesPaynter, just doing some final checks.

- In the Statement of Need in the paper, can you explain what FITS and BATSE mean?
- In the first sentence of the PyGRB section, "to" is missing between "aims bring"
- In the beginning sentence of the next paragraph, write out "four" instead of "4"
- The Ashton et al. reference looks like it is missing the journal name

Actually, it looks like many/most of the references are missing the journal name. I think the problem is that your `.bib` file uses `\apj` and `\aap` for these, but that won't work with our system.
