+++ joss_paper/paper.bib
@@ -3,7 +3,7 @@
          {Scargle}, J.~D. and {Kouveliotou}, C. and {Paciesas}, W.~S. and
          {Meegan}, C.~A. and {Fishman}, G.~J.},
         title = "{Attributes of Pulses in Long Bright Gamma-Ray Bursts}",
-      journal = {\apj},
+      journal = {Astrophysical Journal},
      keywords = {GAMMA RAYS: BURSTS, METHODS: DATA ANALYSIS},
          year = 1996,
         month = mar,
@@ -18,7 +18,7 @@
        author = {{Norris}, J.~P. and {Bonnell}, J.~T. and {Kazanas}, D. and
          {Scargle}, J.~D. and {Hakkila}, J. and {Giblin}, T.~W.},
         title = "{Long-Lag, Wide-Pulse Gamma-Ray Bursts}",
-      journal = {\apj},
+      journal = {Astrophysical Journal},
      keywords = {Gamma Rays: Bursts, ISM: Jets and Outflows, Astrophysics},
          year = 2005,
         month = jul,
@@ -36,7 +36,7 @@ archivePrefix = {arXiv},
 @article{astropy,
     author = {{Astropy Collaboration}},
     title = "{Astropy: A community Python package for astronomy}",
-    journal = {\aap},
+    journal = {Astronomy and Astrophysics},
     archivePrefix = "arXiv",
     eprint = {1307.6212},
     primaryClass = "astro-ph.IM",
@@ -73,7 +73,7 @@ archivePrefix = {arXiv},
          {Payne}, Ethan and {Pitkin}, Matthew D. and {Powell}, Jade and
          {Sarin}, Nikhil and {Smith}, Rory J.~E. and {Thrane}, Eric},
         title = "{BILBY: A User-friendly Bayesian Inference Library for Gravitational-wave Astronomy}",
-      journal = {\apjs},
+      journal = {Astrophysical Journal, Supplement},
      keywords = {gravitational waves, methods: data analysis, methods: statistical, stars: black holes, stars: neutron, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - High Energy Astrophysical Phenomena, General Relativity and Quantum Cosmology},
          year = 2019,
         month = apr,
@@ -92,7 +92,7 @@ archivePrefix = {arXiv},
 @ARTICLE{dynesty,
        author = {{Speagle}, Joshua S.},
         title = "{DYNESTY: a dynamic nested sampling package for estimating Bayesian posteriors and evidences}",
-      journal = {\mnras},
+      journal = {Monthly Notices of the RAS},
      keywords = {methods: data analysis, methods: statistical, Astrophysics - Instrumentation and Methods for Astrophysics, Statistics - Computation},
          year = 2020,
         month = apr,
@@ -119,7 +119,7 @@ archivePrefix = {arXiv},
 @ARTICLE{multinest,
        author = {{Feroz}, F. and {Hobson}, M.~P. and {Bridges}, M.},
         title = "{MULTINEST: an efficient and robust Bayesian inference tool for cosmology and particle physics}",
-      journal = {\mnras},
+      journal = {Monthly Notices of the RAS},
      keywords = {methods: data analysis, methods: statistical, Astrophysics},
          year = 2009,
         month = oct,
@@ -153,7 +153,7 @@ archivePrefix = {arXiv},
 @ARTICLE{feroz,
        author = {{Feroz}, F. and {Hobson}, M.~P.},
         title = "{Multimodal nested sampling: an efficient and robust alternative to Markov Chain Monte Carlo methods for astronomical data analyses}",
-      journal = {\mnras},
[…]
