### 2020-09-18 #2536 @xuanxu
[FITS](https://fits.gsfc.nasa.gov/) is a default image format for astrophysics catalogs, widely used in astronomy so I don't think there's need to explain it.

+1 to all other changes
