### 2020-09-20 c68935f
update to JOSS standards
