+++ joss_paper/paper.md
@@ -33,7 +33,7 @@ $$
 # Statement of Need
 
 The analysis of large amounts of light-curves requires downloading FITS files from the relevant server.
-To do this by hand is a tiring prospect for the 2,704 BATSE GRBs.
+To do this by hand is a tiring prospect for the 2,704 Burst and Transient Source Explorer (BATSE [@batse]) GRBs.
 Having downloaded a FITS file, a scientist would then need to unpack the data from the file, and extract the relevant tables to construct a light-curve.
 They may then want to plot the light-curve for publication, requiring them to write more software to appropriately represent the data.
 Ultimately, they may want to look at population statistics, or compare different GRB pulses.
@@ -43,11 +43,11 @@ This is where `PyGRB` comes in.
 
 # PyGRB
 
-`PyGRB` is a pure Python, open source pulse-fitting package which aims bring gamma-ray burst light-curve fitting and analysis into the 21st century.
-`PyGRB` is able to download the pre-binned BATSE [@batse] light curves (``bfits`` files), in addition to ``tte_list`` time-tagged photon arrival times.
+`PyGRB` is a pure Python, open source pulse-fitting package which aims to bring gamma-ray burst light-curve fitting and analysis into the 21st century.
+`PyGRB` is able to download the pre-binned BATSE light curves (``bfits`` files), in addition to ``tte_list`` time-tagged photon arrival times.
 FITS I/O functionality is provided by `Astropy` [@astropy]. `PyGRB` is built on top of the `Bilby` Bayesian inference library [@bilby], through which `PyGRB` utilises the `Dynesty` [@dynesty] and `Nestle` [@nestle] nested sampling packages [@skilling; @feroz; @multinest].
 
-`PyGRB` makes visually appealing and scientifically instructive light-curves from the 4 broadband energy channel BATSE data.
+`PyGRB` makes visually appealing and scientifically instructive light-curves from the four broadband energy channel BATSE data.
 The main feature of `PyGRB` is its ability to fit analytic light-curve models to data.
 In particular, Bayesian model selection allows the user to determine the most appropriate pulse parameterisation for a particular burst.
 Available pulse parameterisations are Gaussian pulses, FRED pulses and FRED variations.
