### 2020-09-20 284ed77
remove undefined journal macros
