+++ joss_paper/paper.bib
@@ -3,7 +3,7 @@
          {Scargle}, J.~D. and {Kouveliotou}, C. and {Paciesas}, W.~S. and
          {Meegan}, C.~A. and {Fishman}, G.~J.},
         title = "{Attributes of Pulses in Long Bright Gamma-Ray Bursts}",
-      journal = {\apj},
+      journal = {Astrophysical Journal},
      keywords = {GAMMA RAYS: BURSTS, METHODS: DATA ANALYSIS},
          year = 1996,
         month = mar,
